Switched the cargo system to use long instead of float.

+++ Assets/Battle/Factions/Faction.cs
@@ -520,7 +520,7 @@ public class Faction : MonoBehaviour, IPositionConfirmer {
 
     public Ship.ShipBlueprint GetTransportBlueprint() {
         return new Ship.ShipBlueprint(factionIndex, Ship.ShipClass.Transport, "Transport", 3000,
-            new List<CargoBay.CargoTypes>() { CargoBay.CargoTypes.Metal }, new List<float>() { 2400 });
+            new List<CargoBay.CargoTypes>() { CargoBay.CargoTypes.Metal }, new List<long>() { 2400 });
     }
 
     public Transform GetShipTransform() {
