+++ Assets/Battle/Units/Stations/StationAI.cs
@@ -7,7 +7,7 @@ public class StationAI : MonoBehaviour {
     protected Station station;
     [SerializeField] protected float waitSpeed;
     [SerializeField] protected float cargoSpeed;
-    [SerializeField] protected float cargoAmmount;
+    [SerializeField] protected long cargoAmmount;
     protected float waitTime;
     protected float cargoTime;
